refactor frontend, move confluence login
