+++ react-fe/src/util/text.js
